% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\name{add_readme}
\alias{add_readme}
\title{Add readme to the project}
\usage{
add_readme(
  project_path,
  dp_title,
  github_repo_url,
  board_params_set_dried,
  creds_set_dried
)
}
\arguments{
\item{project_path}{Path to the project folder}

\item{dp_title}{readme title}

\item{github_repo_url}{github repo url}

\item{board_params_set_dried}{Character representation of the function for
setting board_params. Use \code{fn_dry()} in combination with
\code{dpi::board_params_set_s3} or \code{dpi::board_params_set_labkey}. See example.}

\item{creds_set_dried}{Character representation of the function for setting
creds. Use \code{fn_dry()} in combination with \code{dpi::creds_set_aws} or
\code{dpi::creds_set_labkey}. See example}
}
\description{
Add readme to the project
}
\keyword{internal}
