% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pins_compatibility}
\alias{check_pins_compatibility}
\title{Check pins package compatibility}
\usage{
check_pins_compatibility()
}
\arguments{
\item{is_legacy}{A boolean variable to indicate if the pins version is legacy one}
}
\description{
Check pins package compatibility
}
\keyword{internal}
