% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map.R
\name{dir_ls_tidy}
\alias{dir_ls_tidy}
\title{Get tidy directory list}
\usage{
dir_ls_tidy(current_dir)
}
\arguments{
\item{current_dir}{path to the current directory}
}
\value{
ls_tidy a tibble
}
\description{
Given a directory returns directory list as a custom tidy tibble
}
\keyword{internal}
