% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map.R
\name{dir_process}
\alias{dir_process}
\title{Process current input directory}
\usage{
dir_process(current_dir, junk_path = character(0))
}
\arguments{
\item{current_dir}{path to the current directory}

\item{junk_path}{path to be dropped from the path prefix
it prevents repeating the capture of folder structure already captured}
}
\value{
read_files a list of read contents
}
\description{
Recursively reads and hashes the content of the current
directory returning the content as a list that matches the structure of the
directory
}
\keyword{internal}
