% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map.R
\name{dir_process_zip}
\alias{dir_process_zip}
\title{Process a zip directory}
\usage{
dir_process_zip(zip_dir)
}
\arguments{
\item{zip_dir}{path to the zipped directory}
}
\value{
read_files a list of read contents
}
\description{
This function simply extends the functionality of \code{dir_process}
to zipped directories.
it unzips in temp directory and once unzipped calls \code{dir_process}
}
\keyword{internal}
