% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map.R
\name{dirTree_flatten}
\alias{dirTree_flatten}
\title{flatten a directory tree}
\usage{
dirTree_flatten(read_files)
}
\arguments{
\item{read_files}{out of call to \code{dir_process}}
}
\value{
a one level deep list of data tables with name of each element
capture the folder structure info
}
\description{
recursively traverses list out of \code{dir_process} and flattens the
list of read data
capturing the structure as path name
}
\keyword{internal}
