% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_commit.R
\name{dp_commit}
\alias{dp_commit}
\title{Commit data product}
\usage{
dp_commit(project_path = fs::path_wd(), commit_description)
}
\arguments{
\item{project_path}{Path to the project folder (default is working directory)}

\item{commit_description}{Commit message}
}
\value{
repo object post commit
}
\description{
Runs a number of checks and commits data product changes
}
\examples{
\dontrun{
dp_commit(commit_description = "First dp commit")
}
}
