% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\name{dp_git_init}
\alias{dp_git_init}
\title{dp git initialization}
\usage{
dp_git_init(
  project_path,
  project_name,
  branch_name,
  github_repo_url,
  board_params_set_dried,
  creds_set_dried,
  git_ignore
)
}
\arguments{
\item{project_path}{Path to the project folder}

\item{project_name}{The name of the project. This is typically the name of
the folder where the project is set}

\item{branch_name}{An abbreviation to capture the specific reason for which
data was processed. Example m3cut (as in month 3 data cut)}

\item{github_repo_url}{the https url for the github repo}

\item{board_params_set_dried}{Character representation of the function for
setting board_params. Use \code{fn_dry()} in combination with
\code{dpi::board_params_set_s3}
or \code{dpi::board_params_set_local}.}

\item{creds_set_dried}{Character representation of the function for setting
creds. Use \code{fn_dry()} in combination with \code{dpi::creds_set_aws}.}

\item{git_ignore}{A character vector of the files and directories to be
ignored by git.}
}
\description{
Initializes the git repo according to the requirements of dp
}
\keyword{internal}
