% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pull.R
\name{dp_pull}
\alias{dp_pull}
\title{Pull data product project from a remote repo}
\usage{
dp_pull(remote_url, branch, verbose = F)
}
\arguments{
\item{remote_url}{a url like \verb{https://github.com/teamAccount/me/dp_test1.git}}

\item{branch}{branch name of the data product (same as that of repo)}

\item{verbose}{T/F if TRUE, it'll show cloning progress}
}
\value{
TRUE
}
\description{
This command pulls data product project from a remote repo
}
\examples{
 \dontrun{
Sys.setenv("GITHUB_PAT" = keyring::key_get("GITHUB_PAT"))
dp_clone(remote_url = "https://github.com/teamAccount/me/dp_test1.git",
branch = us001)
}
}
