% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_push.R
\name{dp_push}
\alias{dp_push}
\title{Push data product to remote repo}
\usage{
dp_push(
  project_path = ".",
  remote_alias = character(0),
  remote_url = character(0)
)
}
\arguments{
\item{project_path}{path to the dp_project (default is current directory)}

\item{remote_alias}{use only to overwrite default alias, or if no default
remote alias has been configured (e.g. legacy).}

\item{remote_url}{use only to overwrite default URL, or if no default remote
URL has been configured (e.g. legacy).}
}
\value{
TRUE
}
\description{
This command pushes the data product to the remote
}
\examples{
\dontrun{
dp_push(project_path = ".")
}
}
