% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{dp_repository_check}
\alias{dp_repository_check}
\title{Check dp repository}
\usage{
dp_repository_check(path)
}
\arguments{
\item{path}{Path to be evaluated}
}
\value{
list of T/F per check
}
\description{
Runs a number of checks and returns a list of T/F per check
}
\keyword{internal}
