% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_structure.R
\name{dp_structure}
\alias{dp_structure}
\title{Structure a data product}
\usage{
dp_structure(data_files_read, config, output = list(), metadata = list())
}
\arguments{
\item{data_files_read}{object generated from \code{dpbuild::dpinput_read()}
containing links to input data}

\item{config}{data product config file from \code{dpbuild::dpconf_get}}

\item{output}{a list of content to be structured under output}

\item{metadata}{a list of content to be structured under metadata}
}
\value{
a list containing README, raw_input, input, and output
}
\description{
This function assembles the data product and properly assigns
class and attributes. For examples, use \code{dpbuild::dpcode_add} to generate a dp_make.R file
}
