% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_tolink.R
\name{dp_tolink}
\alias{dp_tolink}
\title{Converts a data product to a link}
\usage{
dp_tolink(dp_name, dp_version)
}
\arguments{
\item{dp_name}{name of the data product}

\item{dp_version}{version of the data product}
}
\description{
Given the name and version of a data product, it returns a
custom function for retrieving that data product
}
\examples{
\dontrun{
dp_cars_lnk <- dp_tolink(dp_name = "dp-cars-us001", dp_version = "4dc379a")
}
}
