% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_write.R
\name{dp_write}
\alias{dp_write}
\title{Write data product}
\usage{
dp_write(data_object, project_path = ".")
}
\arguments{
\item{data_object}{data_object generated from \code{dp_structure}}

\item{project_path}{path to the project (default is current directory)}
}
\value{
TRUE
}
\description{
This function writes the data product and logs it in \code{.daap/daap_log.yaml}
}
