% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcode_add.R
\name{dpcode_add}
\alias{dpcode_add}
\title{Adds script templates}
\usage{
dpcode_add(project_path, use_targets = T)
}
\arguments{
\item{project_path}{Path to the project}

\item{use_targets}{T/F when TRUE, uses targets instead of drake (recommended)}
}
\value{
repo
}
\description{
Adds dp_make.R and dp_journal.RMD to the dp project repo
}
