% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_get.R
\name{dpconf_get}
\alias{dpconf_get}
\title{Get daap config}
\usage{
dpconf_get(project_path)
}
\arguments{
\item{project_path}{path to project folder}
}
\value{
a list dpconf
}
\description{
Reads in and returns daap config file hydrating any expression
in the process.
}
\details{
This function reads in the yaml config as a list. in the process, it
hydrate any expression for \code{board_params} and \code{creds}. Make sure, environment
variables declared in dried functions are set prior to calling \code{dpconf_get}.
}
