% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\name{dpconf_init}
\alias{dpconf_init}
\title{Initialize daap configuration file}
\usage{
dpconf_init(
  project_path,
  project_name,
  project_description = character(0),
  branch_name,
  branch_description = character(0),
  readme_general_note = character(0),
  board_params_set_dried,
  creds_set_dried,
  is_legacy,
  ...
)
}
\arguments{
\item{project_path}{path to the project folder}

\item{project_name}{the name of the project. This is typically the name of
the folder where the project is set}

\item{project_description}{A high level description of the project. Example:
integrated, clinical and translational data from study x.}

\item{branch_name}{An abbreviation to capture the specific reason for which
data was processed. Example m3cut (as in month 3 data cut)}

\item{branch_description}{A high level description of the branch}

\item{readme_general_note}{Optional general note which will be added as
metadata to the data object}

\item{board_params_set_dried}{Character representation of the function for
setting board_params. Use \code{fn_dry()} in combination with
\code{board_params_set_s3()}, \code{board_params_set_labkey()}, or \code{board_params_set_local()}.}

\item{creds_set_dried}{Character representation of the function for setting
creds. Use \code{fn_dry()} in combination with \code{creds_set_aws()} or \code{creds_set_labkey()}.}

\item{is_legacy}{if pins version is a legacy one (Boolean)}

\item{...}{any other metadata to be captured in the config file}
}
\value{
dpconf
}
\description{
Initializes daap configuration file \code{.daap/daap_config.yaml}
}
\keyword{internal}
