% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_get.R
\name{dpconf_read}
\alias{dpconf_read}
\title{Read daap config}
\usage{
dpconf_read(project_path)
}
\arguments{
\item{project_path}{path to project folder}
}
\value{
a list of raw dpconf where any function call remains dry
}
\description{
Read daap_config.yaml
}
\keyword{internal}
