% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_update.R
\name{dpconf_update}
\alias{dpconf_update}
\title{Modify data product configuration}
\usage{
dpconf_update(
  project_path = fs::path_wd(),
  project_description = character(0),
  branch_name = character(0),
  branch_description = character(0),
  readme_general_note = character(0),
  board_params_set_dried = character(0),
  creds_set_dried = character(0),
  commit_description = "dp_conf modified",
  git_ignore = character(0),
  ...
)
}
\arguments{
\item{project_path}{path to the project folder. The folder name will be used
as project name. If the path doesn't exist it will be created.}

\item{project_description}{A high level description of the project. Example:
integrated, clinical and translational data from study x.}

\item{branch_name}{An abbreviation to capture the specific reason for which
data was processed. Example m3cut (as in month 3 data cut)}

\item{branch_description}{A high level description of the branch}

\item{readme_general_note}{Optional general note which will be added as
metadata to the data object}

\item{board_params_set_dried}{Character representation of the function for
setting board_params. Use \code{fn_dry()} in combination with
\code{board_params_set_s3()}, \code{board_params_set_labkey()}, or \code{board_params_set_local()}.}

\item{creds_set_dried}{Character representation of the function for setting
creds. Use \code{fn_dry()} in combination with \code{creds_set_aws} or \code{creds_set_labkey()}.}

\item{commit_description}{A character string that describes what was updated.}

\item{git_ignore}{a character vector of the files and directories to be added
to existing list ignored by git.}

\item{...}{any other metadata to be captured in the config file}
}
\value{
project path
}
\description{
Manages all modification to data product project configuration
subsequent to the initialization via \code{dp_init}.
}
