% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_get.R
\name{dpconf_validate}
\alias{dpconf_validate}
\title{validate dpconf}
\usage{
dpconf_validate(dpconf, project_path)
}
\arguments{
\item{dpconf}{a dp config}

\item{project_path}{path to project}
}
\description{
This function errors if invalid conf
}
\keyword{internal}
