% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_get.R
\name{dpconf_write}
\alias{dpconf_write}
\title{Write daap config}
\usage{
dpconf_write(project_path, dpconf)
}
\arguments{
\item{project_path}{path to project folder}
}
\value{
a list dpconf
}
\description{
Write in and returns daap config file
}
\keyword{internal}
