% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map0.R
\name{dpinput_flatten}
\alias{dpinput_flatten}
\title{flatten the read dpinput in the format of dp_map}
\usage{
dpinput_flatten(read_dpinput0)
}
\arguments{
\item{read_dpinput0}{out of call to \code{dir_process}}
}
\value{
a one level deep list of data tables with name of each element
capture the folder structure info
}
\description{
recursively traverses list out of \code{dir_process} and flattens the
list of read data
capturing the structure as path name
}
\keyword{internal}
