% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_write.R
\name{dpinput_make}
\alias{dpinput_make}
\title{make data product input manifest}
\usage{
dpinput_make(input_d)
}
\arguments{
\item{input_d}{a list with synced_input elements structured as desired within
a list}
}
\value{
TRUE
}
\description{
Converts structured input into a yaml manifest
}
\keyword{internal}
