% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map.R
\name{dpinput_map}
\alias{dpinput_map}
\title{Map data input into data product workflow}
\usage{
dpinput_map(project_path)
}
\arguments{
\item{project_path}{path to the project}
}
\value{
input_map a data.frame that maps the content of the input_files to be
evaluated
}
\description{
This is helper function to help manage the context of
input_files folder
}
