% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_map0.R
\name{dpinput_map0}
\alias{dpinput_map0}
\title{Get last dpinput}
\usage{
dpinput_map0(project_path = ".")
}
\arguments{
\item{project_path}{current project path}
}
\value{
input_map0 a data.frame that maps the content of the already synced
input content or NULL if no daap_input.yaml
}
\description{
Build the last dpinput from daap_inpt.yaml recorded
}
\keyword{internal}
