% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_read.R
\name{dpinput_read}
\alias{dpinput_read}
\title{Read data product input manifest}
\usage{
dpinput_read(
  daap_input_yaml = yaml::read_yaml(file = "./.daap/daap_input.yaml"),
  add_metadata = F
)
}
\arguments{
\item{daap_input_yaml}{daap_input yaml file imported}

\item{add_metadata}{T/F when TRUE dpinput format will include metadata}
}
\value{
daap_input as a structured list of anonymous functions each when
called retrieves the specific data
}
\description{
Reads yaml DaaP input manifest and returns a structured pinlink
}
