% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_write.R
\name{dplognote_get}
\alias{dplognote_get}
\title{Get log note from data object}
\usage{
dplognote_get(data_object, dataobj_path)
}
\arguments{
\item{data_object}{data_object}

\item{dataobj_path}{path to the data object}
}
\value{
log_note
}
\description{
This builds log note
}
\keyword{internal}
