% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dpname_make}
\alias{dpname_make}
\title{Make Data Product Name}
\usage{
dpname_make(project_name, branch_name)
}
\arguments{
\item{project_name}{Project name}

\item{branch_name}{Branch name}
}
\value{
dp_name a character that will be tagged back as attribute to data_object
}
\description{
A helper function that build data product name (i.e. what the pin is going to be called)
}
