% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flname_xos_get}
\alias{flname_xos_get}
\title{Gets cross OS File Name}
\usage{
flname_xos_get(fl, package = "dpbuild")
}
\arguments{
\item{fl}{just the file name e.g. README.RMD}

\item{package}{package name e.g. dpbuild}
}
\description{
It drops extension that can be OS-specific
}
\keyword{internal}
