% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\name{fn_dry}
\alias{fn_dry}
\title{Dry a called function}
\usage{
fn_dry(fn_called)
}
\arguments{
\item{fn_called}{a function called}
}
\value{
character representation of the called function
}
\description{
Make a text representation of a called function call
}
\examples{
\dontrun{
fn_dry(sum(log(1:10)))
}
}
