% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\name{fn_hydrate}
\alias{fn_hydrate}
\title{Hydrate a dried called function}
\usage{
fn_hydrate(dried_fn)
}
\arguments{
\item{dried_fn}{a function called}
}
\value{
value of the called function given its textual representation
}
\description{
execute and returns the value of function call given its textual
(dried) representation
}
\examples{
\dontrun{
fn_hydrate(fn_dry(sum(log(1:10))))
}
}
\keyword{internal}
