% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_versions}
\alias{get_package_versions}
\title{Gets package versions}
\usage{
get_package_versions(package_names)
}
\arguments{
\item{package_names}{A vector of package names for e.g. c("pins","dpbuild", "dpi", "dpdeploy")}
}
\description{
Gets package versions
}
\keyword{internal}
