% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pin_version}
\alias{get_pin_version}
\title{Get Pins Version Pre Deploy}
\usage{
get_pin_version(d, pin_name, pin_description)
}
\arguments{
\item{d}{data object}

\item{pin_name}{what the pin will be named. For data products, it is encoded in dp_param}

\item{pin_description}{what the pin description will be. For data products, it is encoded in dp_params}
}
\value{
a character version
}
\description{
This get the pins version pre-deploy
}
\keyword{internal}
