% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dppkg_modify.R
\name{get_rspm_pkgmeta}
\alias{get_rspm_pkgmeta}
\title{R the latest package metadata on rspm}
\usage{
get_rspm_pkgmeta(pkg_name, rspm_api_url)
}
\arguments{
\item{pkg_name}{character string name of the package}

\item{rspm_api_url}{the api endpoint of rspm that given pkg_name retrieves
pkg metadata like latest version, remote_sha, etc.}
}
\value{
pkgmeta containing \code{name}, \code{remote_sha}, \code{version}, \code{checksum}, and
\code{repository}
}
\description{
A convenience function to hit the rspm api and get pkg metadata
}
\details{
This is primarily intended to make \code{dppkg_modify} easier to work
with
}
\examples{
 \dontrun{
get_rspm_pkgmeta(pkg_name = "dpi")
}
}
\keyword{internal}
