% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inputmap_clean}
\alias{inputmap_clean}
\title{Clean input_map}
\usage{
inputmap_clean(input_map, remove_id = character(0), force_cleanname = F)
}
\arguments{
\item{input_map}{synced mapped object as returned by \code{dpbuild::dpinput_map}}

\item{remove_id}{a vector of input_data ids to remove. This is for convenience
as setting the input_manifest field \code{to_be_synced} to FALSE can achieve the
same thing. The default value of \code{character(0)} limits removal to any row
with \code{to_be_synced == FALSE}}

\item{force_cleanname}{T/F, if TRUE it ensures each input id name ends up
being unique. If FALSE, it won't clean names unless names are already unique}
}
\value{
input_map pruned and with cleaner names
}
\description{
This function drops unsynced inputs from the input map
and cleans names
}
