% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{is_dp_initiated}
\alias{is_dp_initiated}
\title{Determine if dp repository initiated}
\usage{
is_dp_initiated(path, verbose = F)
}
\arguments{
\item{path}{Path to be evaluated}

\item{verbose}{If TRUE, it will print which tests passed/failed}
}
\value{
TRUE or FALSE
}
\description{
Looks at the path, runs \code{dp_repository_check} and returns TRUE
if any TRUE
}
\details{
All diagnostic tests to check validity of dp repository are run
regardless of choice of checks. Checks determines what subset is considered
in return T/F.
}
\keyword{internal}
