% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{is_valid_dp_repository}
\alias{is_valid_dp_repository}
\title{Determine if valid dp repository}
\usage{
is_valid_dp_repository(
  path,
  checks = c("all", "git", "dp", "renv", "branch"),
  verbose = F
)
}
\arguments{
\item{path}{Path to be evaluated}

\item{checks}{any combination of c("all","git","dp","renv","branch"). default
is all.}

\item{verbose}{If TRUE, it will print which tests passed/failed}
}
\value{
TRUE or FALSE
}
\description{
Looks at the path, runs \code{dp_repository_check} and returns TRUE
if all TRUE
}
\details{
All diagnostic tests to check validity of dp repository are run
regardless of choice of checks. Checks determines what subset is considered
in return T/F. Regardless of the choice of checks, verbose = T
}
