% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_names_codefriendly}
\alias{make_names_codefriendly}
\title{Make names code friendly}
\usage{
make_names_codefriendly(x, make_unique = T)
}
\arguments{
\item{x}{a character string to be converted}

\item{make_unique}{if TRUE it ensures each element of a vector names end up being unique}
}
\value{
the code friendly converted character string
}
\description{
This function tries to provide a more sensible mapping of names to their code friendly version than make.names
}
