% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpinput_read.R
\name{make_pinlink}
\alias{make_pinlink}
\title{Make Pinlink}
\usage{
make_pinlink(synced_input_i)
}
\arguments{
\item{synced_input_i}{a single pin synced data}
}
\value{
a function that receives config and returns the specific data
}
\description{
Converts a synced input into a function that serves as a link.
}
\keyword{internal}
