% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sha1_compatible}
\alias{make_sha1_compatible}
\title{Update exotic column classes of data.frames in a list for sha1 signature}
\usage{
make_sha1_compatible(l)
}
\arguments{
\item{l}{list}
}
\value{
list
}
\description{
This function is to be used before digest::sha1 to handle exotic column classes
}
\keyword{internal}
