% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{purge_local_cache}
\alias{purge_local_cache}
\title{Purge Local Pins Cache}
\usage{
purge_local_cache(path_cache = pins::board_cache_path())
}
\arguments{
\item{path_cache}{path to pins cache. Default is \code{pins::board_cache_path()}}
}
\description{
It completely deletes content of local cache. Use with care!
}
\keyword{internal}
