% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_package_versions_in_config_yaml}
\alias{update_package_versions_in_config_yaml}
\title{Gets package versions}
\usage{
update_package_versions_in_config_yaml(project_path = ".", package_versions)
}
\arguments{
\item{project_path}{path to the dp_project (default is current directory)}

\item{package_versions}{Package version object obtained from get_package_versions method}
}
\description{
Gets package versions
}
\keyword{internal}
