% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pins_compatibility}
\alias{check_pins_compatibility}
\title{Check pins package compatibility}
\usage{
check_pins_compatibility(pins_version = "1.2.0")
}
\arguments{
\item{pins_version}{A string pins version to check against the installed one}
}
\description{
Check pins package compatibility
}
\keyword{internal}
