% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_commit.R
\name{dp_commit}
\alias{dp_commit}
\title{Commit dp}
\usage{
dp_commit(project_path = fs::path_wd(), commit_description)
}
\arguments{
\item{project_path}{Path to the project folder}

\item{commit_description}{commit description}
}
\value{
repo repo object post commit
}
\description{
Runs a number of checks and returns a list of T/F per check
}
