% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pull.R
\name{dp_pull}
\alias{dp_pull}
\title{Pull data product from a remote repo}
\usage{
dp_pull(project_path = fs::path_wd())
}
\arguments{
\item{project_path}{Path to the project folder (default is working directory)}
}
\value{
TRUE
}
\description{
This command pulls the data product from a remote repo
}
\examples{
 \dontrun{
Sys.setenv("GITHUB_PAT" = keyring::key_get("GITHUB_PAT"))
dp_pull()
}
}
