% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pull.R
\name{dp_pull}
\alias{dp_pull}
\title{Pull data product from a remote repo}
\usage{
dp_pull(project_path = ".")
}
\arguments{
\item{project_path}{path to the dp_project}
}
\value{
TRUE
}
\description{
This command pulls the data product from the remote
}
\examples{
 \dontrun{
Sys.setenv("GITHUB_PAT" = keyring::key_get("GITHUB_PAT"))
dp_pull(project_path = ".")
}
}
