% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_structure.R
\name{dp_structure}
\alias{dp_structure}
\title{Structure a data product}
\usage{
dp_structure(data_files_read, config, output = list(), metadata = list())
}
\arguments{
\item{data_files_read}{this is the output of read_data_files}

\item{config}{this is the config file read}

\item{output}{a list of content to be structured under output}

\item{metadata}{a list of content to be structured under metadata}
}
\value{
a list containing README, raw_input, input, and output
}
\description{
Structures data product and properly assigns class and attributes
}
\details{
This function assembles data object
}
