% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpbuild-package.R
\docType{package}
\name{dpbuild-package}
\alias{dpbuild}
\alias{dpbuild-package}
\title{dpbuild: A package to manage building of data products}
\description{
An R package intended to help set up and manage a data product project.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/amashadihossein/dpbuild/issues}
}

}
\author{
\strong{Maintainer}: Afshin Mashadi-Hossein \email{amashadihossein@gmail.com} (\href{https://orcid.org/0000-0002-2337-2195}{ORCID})

}
\keyword{internal}
