% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpcode_add.R
\name{dpcode_add}
\alias{dpcode_add}
\title{Adds script templates}
\usage{
dpcode_add(project_path)
}
\arguments{
\item{project_path}{Path to the Project}
}
\value{
repo
}
\description{
Adds dp_assembler.R and dp_structure.R to the dp project repo
}
