% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpconf_get.R
\name{dpconf_get}
\alias{dpconf_get}
\title{Get data product config information}
\usage{
dpconf_get(project_path = ".")
}
\arguments{
\item{project_path}{path to project folder}
}
\value{
a list dpconf
}
\description{
Reads in data product config file \code{.daap/daap_config.yaml}and
returns config details
}
\details{
This function reads in the yaml config as a list. In the process, it
hydrates any expression for \code{board_params} and \code{creds}. Make sure environment
variables declared in dried functions are set prior to calling \code{dpconf_get}.
}
