% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inputmap_clean}
\alias{inputmap_clean}
\title{Clean input_map}
\usage{
inputmap_clean(input_map, force_cleanname = F)
}
\arguments{
\item{input_map}{synced mapped object as returned by \code{dpbuild::dpinput_map}}

\item{force_cleanname}{if TRUE it ensures each element of a vector names end up being unique. If not, it won't clean names unless clean is also unique}
}
\value{
input_map pruned and with cleaner names
}
\description{
This function drops unsynced or not-to-be-synced and pos
}
