% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{is_dp_initiated}
\alias{is_dp_initiated}
\title{Determine if dp repository initiated}
\usage{
is_dp_initiated(path, checks = c("git", "dp", "renv"), verbose = F)
}
\arguments{
\item{path}{Path to be evaluated}

\item{checks}{any combination of c("git","dp","renv"). default
is any.}

\item{verbose}{If TRUE, it will print which tests passed/failed}
}
\value{
TRUE or FALSE
}
\description{
Looks at the path, runs \code{dp_repository_check} and returns TRUE
if all TRUE
}
\details{
Any diagnostic tests to check validity of dp repository are run
regardless of choice of checks. Checks determines what subset is considered
in return T/F. Regardless of the choice of checks, verbose = F
}
\keyword{internal}
