% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readme_get}
\alias{readme_get}
\title{Get Readme to be appended to the data object}
\usage{
readme_get(d, general_note)
}
\arguments{
\item{d}{data object}

\item{general_note}{a character string to be added to general notes for this data object branch/commit}
}
\value{
readme text
}
\description{
This function builds readme metadata
}
