% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_init.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{board_params_set_s3}
\alias{board_params_set_local}
\alias{creds_set_aws}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dpi}{\code{\link[dpi]{board_params_set_local}}, \code{\link[dpi]{board_params_set_s3}}, \code{\link[dpi]{creds_set_aws}}}
}}

